Incorporate Lucy and Joss's feedback.
Groups of 3. An odd number per group ensures that the same person in
each group isn't writing all the tests.

Use Joss's suggestions for reflection question wording.

Make the use of "I" and "the mentor" consistent.

Make it clear that we will stand up and move away from computers when we
need a break to think and discuss the solution. The glow of a computer
saps creativity and ability to think, IMHO. And standing up may improve
circulation also.